refactor: Remove cpy dependency, add tests
The cpy dependency, though lovely, is a little slow on adopting Node v10, and we aren't using much
of its API surface. I was, and still am, a little hesitant to make any major changes in this
project, and this is one - it removes the cpy dependency, and instead DIYs the filesystem changes.

Fixes #136