refactor: Remove audio support (#191)
BREAKING CHANGE: Removes audio support

In an effort to refocus big, I'm removing audio support. It was a cool
feature, and will continue to be available in older versions of big. If
you'd like to keep it going, feel free to fork and maintain a version
with the feature. According to the poll (#188), usage of this feature
was minimal, and it occupied a decent chunk of big's surface area.