refactor: Switch from Rubik to system font stack (#194)
Big started with Helvetica, then switched to Rubik, and is now going back to
system fonts, because options like San Francisco are really good. Fonts are
still easily configurable if you want.

This simplifies usage because presentations will include fewer files, it
saves on CSS, and will make presentations faster because browsers will
not need to load and parse the web fonts.